docker-compose.yml: use proper timezone
As timezone-naive datetime.today() instead of datetime.now(TZ) is used
in the code the application has to guess the timestamp. The easiest
workaround is to set proper timezone in the app container.

Setting open in the postgres won't affect anything, so not defining TZ and
PGTZ for the db container.